+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.6.0
+
+* added an info line
+
+## 0.5.0
+
+* added loop to retry initial ssh session on Linode
+
+## 0.4.0
+
+* set default ssh timeout to 600 seconds
+* added some more information lines
+
 ## 0.3.0
 
 * added option for ssh timeout, set default to 60 seconds